refactor(core): Minor leftover cleanups in native Python runner
