+++ packages/@n8n/task-runner-python/src/task_executor.py
@@ -26,12 +26,12 @@ from src.config.security_config import SecurityConfig
 
 from src.message_types.broker import NodeMode, Items
 from src.message_types.pipe import (
-    PipeMessage,
     PipeResultMessage,
     PipeErrorMessage,
     TaskErrorInfo,
     PrintArgs,
 )
+from src.pipe_reader import PipeReader
 from src.constants import (
     EXECUTOR_CIRCULAR_REFERENCE_KEY,
     EXECUTOR_USER_OUTPUT_KEY,
@@ -100,7 +100,7 @@ class TaskExecutor:
 
         print_args: PrintArgs = []
 
-        pipe_reader = PipeReaderThread(read_conn.fileno(), read_conn)
+        pipe_reader = PipeReader(read_conn.fileno(), read_conn)
         pipe_reader.start()
 
         try:
@@ -496,23 +496,6 @@ class TaskExecutor:
     # ========== pipe I/O ==========
 
     @staticmethod
-    def _read_exact_bytes(fd: int, n: int) -> bytes:
-        """Read exactly n bytes from file descriptor.
-
-        Uses os.read() instead of Connection.recv() because recv() pickles.
-        Preallocates bytearray to avoid repeated reallocation.
-        """
-        result = bytearray(n)
-        offset = 0
-        while offset < n:
-            chunk = os.read(fd, n - offset)
-            if not chunk:
-                raise EOFError("Pipe closed before reading all data")
-            result[offset : offset + len(chunk)] = chunk
-            offset += len(chunk)
-        return bytes(result)
-
-    @staticmethod
     def _write_bytes(fd: int, data: bytes):
         total_written = 0
         while total_written < len(data):
@@ -520,59 +503,3 @@ class TaskExecutor:
             if written == 0:
                 raise OSError("Write failed")
             total_written += written
-
-
-class PipeReaderThread(threading.Thread):
-    """Background thread that reads result from pipe."""
-
-    def __init__(self, read_fd: int, read_conn: PipeConnection):
-        super().__init__()
-        self.read_fd = read_fd
-        self.read_conn = read_conn
-        self.pipe_message: PipeMessage | None = None
-        self.message_size: int | None = None  # bytes
-        self.error: Exception | None = None
-
-    def run(self):
-        try:
-            length_bytes = TaskExecutor._read_exact_bytes(
-                self.read_fd, PIPE_MSG_PREFIX_LENGTH
-            )
-            length_int = int.from_bytes(length_bytes, "big")
-            if length_int <= 0:
-                raise InvalidPipeMsgLengthError(length_int)
-            self.message_size = length_int
-            data = TaskExecutor._read_exact_bytes(self.read_fd, length_int)
-            parsed_msg = json.loads(data.decode("utf-8"))
-            self.pipe_message = self._validate_pipe_message(parsed_msg)
-        except Exception as e:
-            self.error = e
-        finally:
-            self.read_conn.close()
-
-    def _validate_pipe_message(self, msg) -> PipeMessage:
-        if not isinstance(msg, dict):
-            raise InvalidPipeMsgContentError(f"Expected dict, got {type(msg).__name__}")
-
-        if "print_args" not in msg:
-            raise InvalidPipeMsgContentError("Message missing 'print_args' key")
-
-        if not isinstance(msg["print_args"], list):
-            raise InvalidPipeMsgContentError("'print_args' must be a list")
-
-        has_result = "result" in msg
-        has_error = "error" in msg
-
-        if not has_result and not has_error:
-            raise InvalidPipeMsgContentError("Msg is missing 'result' or 'error' key")
-
-        if has_result and has_error:
-            raise InvalidPipeMsgContentError("Msg has both 'result' and 'error' keys")
-
-        if has_result and not isinstance(msg["result"], list):
-            raise InvalidPipeMsgContentError("'result' must be a list")
-
-        if has_error and not isinstance(msg["error"], dict):
-            raise InvalidPipeMsgContentError("'error' must be a dict")
-
-        return cast(PipeMessage, msg)
